feat: MapGrid x, y Int -> Double 수정

+++ modules/core/src/main/kotlin/im/etap/forecast/core/map/MapGrid.kt
@@ -1,3 +1,3 @@
 package im.etap.forecast.core.map
 
-data class MapGrid(val x: Int, val y: Int)
+data class MapGrid(val x: Double, val y: Double)
